Don't raise an error when runresult failed to compile.
The Meson documentation states that the values of runresult's
returncode(), stdout(), and stderr() methods should be considered
undefined if both the runresult was generated by a compiler's run()
method and the compilation itself was unsuccessful.

Currently, muon raises an error and stops interpreting the build
script when the user invokes any of these functions in this case.

This behaviour diverges from Meson, which does not raise an error
in this case, but rather sets these values of be undefined, for
some value of 'undefined'.

Harmonizing muon's behaviour with Meson's allows Mesa's build
to reach i18n.

I'm not too sure about this change. It seems semantically quite
impactful and not in a good way.